+++ CHANGELOG.md
@@ -12,7 +12,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 - Temperature dependent fluid properties (issue #143).
 - Freezing point to fluid data (issue #314).
 - Mass and volume percentage support(issue #350).
-- Add Thermox DTX and Coolflow NFT
+- Add commercial fluids: Thermox DTX and Coolflow NTP from Hydratech.
 
 ### Changed